Gewichtung double

+++ code/src/abicalc/Fach.java
@@ -124,7 +124,7 @@ public class Fach extends Component implements java.io.Serializable{//Datenstruk
         button_plus.addActionListener(new java.awt.event.ActionListener() {		//Code fürs Noten hinzufügen
 	        public void actionPerformed(java.awt.event.ActionEvent e) {
 	          
-	        	Note n = new Note(Integer.parseInt(txt_Gewichtung.getText()), txt_Name.getText(), Integer.parseInt(txt_Note.getText()), panel_scrollContent);
+	        	Note n = new Note(Double.parseDouble(txt_Gewichtung.getText()), txt_Name.getText(), Integer.parseInt(txt_Note.getText()), panel_scrollContent);
 	        	panel_scrollContent.add(n);
 	        	notenliste.add(n);
 	        	main.validate();
